Atencioon al cliente 99% done
Falta poner que cuando el trabajador sale el cliente asignado a su puesto sea redirigido a la cola de espera en el primero lugar

+++ Obligatorio/Obligatorio2/src/modelo/Puesto.java
@@ -74,8 +74,8 @@ public class Puesto extends Observable{
         
     }
     
-    public float tiempoPromedio() {
-        float ret = 0;
+    public int tiempoPromedio() {
+        int ret = 0;
         for (NumeroAtencion np : numerosProcesados) {
             ret += np.duracionAten();
         }
